feat: improve `enforceCapability` type inference
This solution still requires work, to avoid the need for any explicit typing, but it is now possible to access validated payloads in the `validate` guard function.

This was accomplished by making `enforceCapability` generic over validated parameters so that you can specify the expected `json`/`param`/`query` results from validator calls in the preceding middleware.