refactor(tests): update fixture setup to use unified context object and add assertion helpers
Because tests are async, import `expect` direct from vitest means expect failures are not necessarily associated to the failed test case making debugging difficult. This change unifies the context object, and then passes that object to custom assertions so that if there is a failure it should be associated to the correct test case.

This also updates `assertDefined` to explicitly check for `null` which was previously missed.

Improve logging in `expectSuccessResponse` and `expectFailureResponse`, however, a wider refactor of `test-client.ts` is needed here to better surface api failures.

Added `expectAccount` and `assertDocumentCount` assertion helpers to reduce boilerplate.